#46 refactor : Shop 비교로직 중복제거 및 공통 응답 DTO로 수정
